appbuilder-qa-chat/requirements.txt
appbuilder-sdk
chainlit